refactor(react): convert `Search` to TypeScript
Closes #12553
